requirements.txt
anyio==4.6.2.post1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
python-telegram-bot==21.7
requests==2.32.3
schedule==1.2.2
sniffio==1.3.1
soupsieve==2.6
urllib3==2.2.3
